LayoutTests/imported/w3c/web-platform-tests/navigation-api/currententrychange-event/navigation-back-forward-same-doc-expected.txt


FAIL currententrychange fires for same-document navigation.back() and navigation.forward() assert_equals: expected 3 but got 2
